move to consensus-types folder
